Move NormType to lapack_traits
